fix: printer parsing crashing on certain systems (#117)
- Use the same version of pdf-to-printer (newer versions have an error
in the validation of the output and crash) as the installer.

- Improve logger for the printer service to print the actual error if no
stacktrace is available.

Note: this is meant to be a quick fix, the package pdf-to-printer is
archived currently, we might need to fork it or use another library in
the future. Also the logging config should be looked at in the future as
this error was not visible without printing it directly.